Extract generic form wizard logic out from create journey
This change refactors the base create journey controller to extract
logic that is not create specific into a separate controller so that
it can be re-used for other form wizard journeys.
