+++ app/move/controllers/create/index.js
@@ -1,12 +1,12 @@
 const Assessment = require('./assessment')
-const Form = require('./form')
+const Base = require('./base')
 const MoveDetails = require('./move-details')
 const PersonalDetails = require('./personal-details')
 const Save = require('./save')
 
 module.exports = {
   Assessment,
-  Form,
+  Base,
   MoveDetails,
   PersonalDetails,
   Save,
