+++ app/move/controllers/create/move-details.js
@@ -1,12 +1,12 @@
 const dateFns = require('date-fns')
 
-const FormController = require('./form')
+const CreateBaseController = require('./base')
 const filters = require('../../../../config/nunjucks/filters')
 const fieldHelpers = require('../../../../common/helpers/field')
 const referenceDataService = require('../../../../common/services/reference-data')
 const referenceDataHelpers = require('../../../../common/helpers/reference-data')
 
-class MoveDetailsController extends FormController {
+class MoveDetailsController extends CreateBaseController {
   async configure(req, res, next) {
     try {
       const courts = await referenceDataService.getLocations({
